requirements.txt
Django==4.2.13
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg2==2.9.3
djangorestframework-simplejwt==5.3.1